pyproject.toml
[project]
name = "cabjovi"
version = "0.1.0"
description = "Raspberry Pi music player service with time-scheduled playback and GPIO mute control"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Players",
]
dependencies = [
    "typer~=0.21",
    "gpiod~=2.4",
    "pyalsaaudio~=0.11",
]

[project.scripts]
cabjovi = "cabjovi.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
